Basic inspector showing entity information panel
Took 50 minutes
