+++ crates/dream-editor/Cargo.toml
@@ -9,6 +9,7 @@ crate-type = ["cdylib", "rlib"]
 [dependencies]
 egui-winit = { version = "0.22.0", git = "https://github.com/emilk/egui.git", default-features = false, features = ["links", "wayland"] }
 egui-wgpu = { version = "0.22.0", git = "https://github.com/emilk/egui.git" }
+crossbeam-channel = { workspace = true }
 cfg-if = { workspace = true }
 winit = { workspace = true }
 cgmath = { workspace = true }
